Write tests for if the server is full, reject the client
