+++ src/client/network/client.cpp
@@ -70,6 +70,7 @@ void Client::handlePacket(ClientPacket& packet)
     // clang-format off
     switch (packet.command) {
         case Cmd::HandshakeChallenge: onHandshakeChallenge(packet); break;
+        case Cmd::ConnectionAcceptance: onConnectionAcceptance(packet); break;
         default: break;
     }
     // clang-format on
